Save whitelist status of all apps.
When app is asked to be whitelisted (or removed from the whitelist),
NMPS checks if the app has internet permission, since if it doesn't the
firewall rules do not need to be changed.

But even if the optimization kicks in, it should still mark the app as
whitelisted, otherwise it will leave to inconsistencies like the UI
never displaying it as whitelisted.

BUG: 26685616
Change-Id: I6e287092d7bf8b7aa3a1bda707b58b128a22bffa