Make sure that Toast is always shown by Meta-Space.
It turns out that my previous CL [1] does not always show Toast when
Meta-Space is hit multiple times in a short period.  With this CL, we
attempt to make sure that the Toast is shown every time when Meta-Space
rotates the current IME subtype.

This CL also removes a rule that we do not show Toast when the IME
window has InputMethodService.IME_VISIBLE bit as it turns out to be
confusing.  In Android N, we just provide a consistent behavior
no matter the IME window is visible or not.

 [1]: If3fe17adbffe4c1125783fc77fed3cfe78fc7933
      f68cdc0152d4e20d4c3595d5c78b476cbd5a7d32

Bug: 27547054
Bug: 28127795
Change-Id: I551c51cd672dba6e7693d23cb181007f51b0de9d