Don't run the WebView preparation if using the same non-replaced package
Whenever all primary WebView packages become disabled or uninstalled the
fallback will be enabled and then used. The enabling causes an intent to
be broadcasted and that intent then causes WebViewUpdateService to rerun
its WebView preparation phase.
With this CL we ignore intents for changing the enabled-state of the
current WebView package if we will still use the same WebView package.
The only time we should rerun the preparation logic when the package
is the same (if it has the same name) is if it has been replaced by a
new version.

Bug: 27899444
Change-Id: Iea029e9f854e31b25b1803048c470fb6732e5d0f