docs: Updated support library overview, CPd from mnc-mr-docs
Started this CL when mnc-mr-docs was the active branch, but now
the production builds come from mnc-docs.

Per feedback from the support library team, updated the support
library overview page to better explain what the libraries are
and what they're for. This is the beginning of a big overhaul of the
support library docs.

The "overview" section has been rewritten and expanded in this CL.
The "revisions" section hasn't been touched, except that the heading
was changed to "Support Library Revision History". (I'd like to
move "Revision History" to a separate page, but I'll do it in a
follow-on CL.)

See first comment for doc stage location.

bug: 27274455
Change-Id: I12ce21c948421e1bd0be5d3151845f900750e7b1
(cherry picked from commit 623060a4ba5aaf3d6c27b88eec544ef9c1c27961)