Respect WRAP_CONTENT when resolving drop down position.
We recently added the width/height <0 check to prevent
passing unresolved widths in to tryFitVertical or
tryFitHorizontal and produce strange behaviors. This made
more sense than the code before it but still not that much
sense. We need to treat MATCH_PARENT and WRAP_CONTENT differently
as the documentation suggests.

Bug: 28267951
Change-Id: I6f2f248c8d1f63d31bdd75c24f7250c28381c712