Fixed an issue where children would animate strangely when removed
Previously we tried to fix a notification flicker when the summary
of a group was removed. This lead to a few issues that group children
would now also generate an animation and the whole notification would
just look weird.
We are now cancelling notification children as soon as their parent
is cancelled already in systemUI to ensure that the animations are
properly rendered.

Change-Id: Ie639b4ad28bdb55d922308e04c14a4e5b32b90bb
Fixes: 28190616