Fixed bad focus stack state when focused docked stack is removed
ag/990438 allowed for stack focus not to be adjusted when the currently
focused stack is removed since the caller will be moving the focus to
the right stack next. This was incorrectly enabled for the positionTask
call were the stack will not be moved to the front since the functionality
is used to move task within stacks we don't want to gain focus.

Bug: 28616824
Change-Id: Ia6aea17c0335d036a17b83fed28799428ca58eda